Fixed the weird bug, now just figuring out type
