+++ src/components/fields/ColorSelector/colors.tsx
@@ -1,0 +1,20 @@
+export const inputClasses: Record<string, string> = {
+  'Ally Green': 'bg-inreach-ally-green',
+  'Light Grey': 'bg-inreach-light-grey',
+  'Soft Black': 'bg-inreach-soft-black',
+  'Dark Grey': 'bg-inreach-dark-grey',
+  'White': 'bg-inreach-white',
+  'Teal': 'bg-inreach-teal',
+  'Cornflower': 'bg-inreach-cornflower',
+  'Cool Grey': 'bg-inreach-cool-grey',
+  'Pink': 'bg-inreach-pink',
+  'Light Blue': 'bg-inreach-light-blue',
+  'Purple': 'bg-inreach-purple',
+  'Dark Blue': 'bg-inreach-dark-blue',
+  'Green': 'bg-inreach-green',
+  'Yellow': 'bg-inreach-yellow',
+  'Orange': 'bg-inreach-orange',
+  'Red': 'bg-inreach-red',
+  'Brown': 'bg-inreach-brown',
+  'Dark Brown': 'bg-inreach-dark-brown',
+}
