+++ src/lib/utils.ts
@@ -4,24 +4,3 @@ import { twMerge } from 'tailwind-merge'
 export function cn(...inputs: ClassValue[]) {
 	return twMerge(clsx(...inputs))
 }
-
-export const inputClasses: Record<string, string> = {
-  'Ally Green': 'bg-inreach-ally-green',
-  'Light Grey': 'bg-inreach-light-grey',
-  'Soft Black': 'bg-inreach-soft-black',
-  'Dark Grey': 'bg-inreach-dark-grey',
-  'White': 'bg-inreach-white',
-  'Teal': 'bg-inreach-teal',
-  'Cornflower': 'bg-inreach-cornflower',
-  'Cool Grey': 'bg-inreach-cool-grey',
-  'Pink': 'bg-inreach-pink',
-  'Light Blue': 'bg-inreach-light-blue',
-  'Purple': 'bg-inreach-purple',
-  'Dark Blue': 'bg-inreach-dark-blue',
-  'Green': 'bg-inreach-green',
-  'Yellow': 'bg-inreach-yellow',
-  'Orange': 'bg-inreach-orange',
-  'Red': 'bg-inreach-red',
-  'Brown': 'bg-inreach-brown',
-  'Dark Brown': 'bg-inreach-dark-brown',
-}
